+++ src/components/basic/StaticTable/StaticTable.stories.tsx
@@ -52,7 +52,7 @@ Table.args = {
           inputValue: 'row1 col2',
           isValid: (value) => console.log('Checking Validation', value),
           errorMessage: 'Please enter valid value.',
-        }
+        },
       ],
       [
         {
@@ -61,8 +61,8 @@ Table.args = {
         {
           editIcon: false,
           clickableLink: 'https://google.com',
-        }
-      ]
+        },
+      ],
     ],
   },
   handleEdit: () => console.log('CLICKED'),