+++ src/components/content/Cards/Card.tsx
@@ -260,7 +260,10 @@ export const Card = ({
                 padding: '45px 90px 10px 90px',
                 '-webkit-transform': 'rotate(45deg) scale(0.75,1)',
                 zIndex: 10,
-                background: subscriptionStatus?.toLowerCase() === 'pending' ? 'linear-gradient(239deg, rgba(255,120,44,1) 0%, rgba(255,179,38,1) 50%)' : '#B3CB2D'
+                background:
+                  subscriptionStatus?.toLowerCase() === 'pending'
+                    ? 'linear-gradient(239deg, rgba(255,120,44,1) 0%, rgba(255,179,38,1) 50%)'
+                    : '#B3CB2D',
               }}
             >
               {subscriptionStatus}