+++ CHANGELOG.md
@@ -4,7 +4,7 @@
   - Add column clickable
   - Add active on hover of edit icon
 
-## 2.0.16
+## 2.0.18
 
 - ImageGallery
   - Use Image type